refactor(stats-box-grid): code-split component and simplify
