+++ src/components/StatsBoxGrid/use-stats-box-grid.tsx
@@ -1,38 +1,27 @@
 import React, { useState, useEffect } from "react"
-import { useIntl } from "react-intl"
 import axios from "axios"
-import { kebabCase } from "lodash"
-import { AreaChart, ResponsiveContainer, Area, XAxis } from "recharts"
+import { useIntl } from "react-intl"
 
-import { VStack, Grid, Box, Icon, Button, Text, Flex } from "@chakra-ui/react"
-import { MdInfoOutline } from "react-icons/md"
-
-import Translation from "./Translation"
-import Tooltip from "./Tooltip"
-import Link from "./Link"
-import StatErrorMessage from "./StatErrorMessage"
-import StatLoadingMessage from "./StatLoadingMessage"
-
+import { RangeSelector } from "./RangeSelector"
+import { Direction } from "../../types"
+import { GATSBY_FUNCTIONS_PATH } from "../../constants"
+import { getData } from "../../utils/cache"
+import { Lang } from "../../utils/languages"
 import {
-  isLangRightToLeft,
-  translateMessageId,
   getLocaleForNumberFormat,
-} from "../utils/translations"
-import { getData } from "../utils/cache"
+  translateMessageId,
+  isLangRightToLeft,
+} from "../../utils/translations"
 
-import { GATSBY_FUNCTIONS_PATH } from "../constants"
-import { Lang } from "../utils/languages"
-import { Direction } from "../types"
+export const ranges = ["30d", "90d"] as const
 
-const ranges = ["30d", "90d"] as const
-
-interface State {
+export interface State {
   value: string
   data: Array<{ timestamp: number }>
   hasError: boolean
 }
 
-interface Metric {
+export interface Metric {
   title: string
   description: string
   state: State
@@ -41,211 +30,6 @@ interface Metric {
   apiUrl: string
   apiProvider: string
 }
-
-interface IGridItemProps {
-  metric: Metric
-  dir?: Direction
-}
-
-const GridItem: React.FC<IGridItemProps> = ({ metric, dir }) => {
-  const { title, description, state, buttonContainer, range } = metric
-  const isLoading = !state.value
-  const value = state.hasError ? (
-    <StatErrorMessage />
-  ) : isLoading ? (
-    <StatLoadingMessage />
-  ) : (
-    <VStack>
-      <Box>
-        {state.value}{" "}
-        <Tooltip content={tooltipContent(metric)}>
-          <Icon
-            as={MdInfoOutline}
-            boxSize={6}
-            fill="text"
-            mr={2}
-            _hover={{ fill: "primary" }}
-            _active={{ fill: "primary" }}
-            _focus={{ fill: "primary" }}
-          ></Icon>
-        </Tooltip>
-      </Box>
-    </VStack>
-  )
-
-  // Returns either 90 or 30-day data range depending on `range` selection
-  const filteredData = (data: Array<{ timestamp: number }>) => {
-    if (!data) return
-    if (range === ranges[1]) return [...data]
-    return data.filter(({ timestamp }) => {
-      const millisecondRange = 1000 * 60 * 60 * 24 * 30
-      const now = new Date().getTime()
-      return timestamp >= now - millisecondRange
-    })
-  }
-
-  const chart: React.ReactNode = (
-    <ResponsiveContainer width="100%" height="100%">
-      <AreaChart
-        data={filteredData(state.data)}
-        margin={{ left: -5, right: -5 }}
-      >
-        <defs>
-          <linearGradient
-            id={`colorUv-${kebabCase(title)}`}
-            x1="0"
-            y1="0"
-            x2="0"
-            y2="1"
-          >
-            <stop offset="5%" stopColor="#8884d8" stopOpacity={1} />
-            <stop offset="95%" stopColor="#8884d8" stopOpacity={0} />
-          </linearGradient>
-          <linearGradient
-            id={`colorPv-${kebabCase(title)}`}
-            x1="0"
-            y1="0"
-            x2="0"
-            y2="1"
-          >
-            <stop offset="5%" stopColor="#82ca9d" stopOpacity={0.8} />
-            <stop offset="95%" stopColor="#82ca9d" stopOpacity={0} />
-          </linearGradient>
-        </defs>
-        <Area
-          type="monotone"
-          dataKey="value"
-          stroke="#8884d8"
-          fillOpacity={0.3}
-          fill={`url(#colorUv-${kebabCase(title)})`}
-          connectNulls={true}
-        />
-        <XAxis dataKey="timestamp" axisLine={false} tick={false} />
-      </AreaChart>
-    </ResponsiveContainer>
-  )
-
-  return (
-    <Flex
-      position="relative"
-      color="text"
-      height={80}
-      flexDirection="column"
-      justifyContent="space-between"
-      alignItems="flex-start"
-      borderX={{
-        base: "0px solid #000000",
-        lg: "1px solid",
-      }}
-      borderY="1px solid"
-      marginTop={{
-        base: "-1px",
-        lg: "0",
-      }}
-      padding={{ base: "2rem 1rem 1rem", lg: "1.5rem" }}
-    >
-      <div>
-        <Text
-          fontSize="xl"
-          mb={2}
-          color="text"
-          textTransform="uppercase"
-          fontFamily="monospace"
-        >
-          {title}
-        </Text>
-        <p>{description}</p>
-      </div>
-      {!state.hasError && !isLoading && (
-        <>
-          <Box
-            position="absolute"
-            left={0}
-            bottom={0}
-            width="100%"
-            height="65%"
-          >
-            {chart}
-          </Box>
-          {dir === "rtl" ? (
-            <Box
-              position="absolute"
-              bottom="20px"
-              fontFamily="monospace"
-              left="20px"
-            >
-              {buttonContainer}
-            </Box>
-          ) : (
-            <Box
-              position="absolute"
-              bottom="20px"
-              fontFamily="monospace"
-              right="20px"
-            >
-              {buttonContainer}
-            </Box>
-          )}
-        </>
-      )}
-      <Text
-        position="absolute"
-        bottom="8%"
-        fontSize={{ base: "max(8.8vw, 48px)", lg: "min(4.4vw, 4rem)" }}
-        fontWeight={600}
-        marginTop={0}
-        marginBottom={4}
-        color="text"
-        flexWrap="wrap"
-        textOverflow="ellipsis"
-      >
-        {value}
-      </Text>
-    </Flex>
-  )
-}
-
-const tooltipContent = (metric: Metric) => (
-  <div>
-    <Translation id="data-provided-by" />{" "}
-    <Link to={metric.apiUrl}>{metric.apiProvider}</Link>
-  </div>
-)
-
-interface IRangeSelectorProps {
-  state: string
-  setState: (state: string) => void
-}
-
-const RangeSelector: React.FC<IRangeSelectorProps> = ({ state, setState }) => (
-  <div>
-    {ranges.map((range, idx) => (
-      <Button
-        onClick={() => setState(ranges[idx])}
-        key={idx}
-        color={""}
-        background="background"
-        fontFamily="monospace"
-        fontSize="xl"
-        padding="2px 15px"
-        borderRadius="1px"
-        border="1px solid"
-        cursor="pointer"
-        _focus={{ outline: "none" }}
-        _hover={{ color: "" }}
-        _active={{ color: "" }}
-        _disabled={{
-          cursor: "default",
-          opacity: "0.7",
-        }}
-        size="sm"
-        backgroundColor={state === ranges[idx] ? "homeBoxPurple" : ""}
-      >
-        {range}
-      </Button>
-    ))}
-  </div>
-)
 
 interface IFetchPriceResponse {
   prices: Array<[number, number]>
@@ -265,9 +49,7 @@ interface IFetchTxResponse {
   transactionCount: number
 }
 
-export interface IProps {}
-
-const StatsBoxGrid: React.FC<IProps> = () => {
+export const useStatsBoxGrid = () => {
   const intl = useIntl()
 
   const [ethPrices, setEthPrices] = useState<State>({
@@ -530,30 +312,10 @@ const StatsBoxGrid: React.FC<IProps> = () => {
       range: selectedRangeNodes,
     },
   ]
-  const dir = isLangRightToLeft(intl.locale as Lang) ? "rtl" : "ltr"
-  return (
-    <Grid
-      display={{
-        base: "flex",
-        lg: "grid",
-      }}
-      gridTemplateColumns="repeat(2, 1fr)"
-      margin={{
-        base: "0",
-        sm: "2rem 0 0",
-        lg: "2rem 2rem 0",
-      }}
-      borderRadius="sm"
-      flexDirection={{
-        base: "column",
-        lg: "column",
-      }}
-    >
-      {metrics.map((metric, idx) => (
-        <GridItem key={idx} metric={metric} dir={dir} />
-      ))}
-    </Grid>
-  )
-}
+  const dir: Direction = isLangRightToLeft(intl.locale as Lang) ? "rtl" : "ltr"
 
-export default StatsBoxGrid
+  return {
+    metrics,
+    dir,
+  }
+}
